moved event list to src
